requirements.txt
streamlit
opencv-python-headless
cmake
dlib
numpy
Pillow
